Move js -> omi-client; move package to top-level
Rename the library from omi-summer-lab to omi-client, and rename the js
directory accordingly.  This change gives the directory more obbious
meaning.  Updates all references as needed.

In order to simplify installation for third party applications, move
the package json up to the root of the repo, such that the omi-client
library can be installed via the github repo URL.

Signed-off-by: Peter Schwarz <pschwarz@bitwise.io>